wip commit
Only use new method on Linux

dummy commit tbh
